Fixed simple alg, also made it check both y and z
The simple algorithm didn't save the best value correctly resulting in the optimizer going in a random direction and it would be 'better'.
This is fixed. Also now the simple alg checks both y and z before making a decision of which way to go, before it would first check z and only check y if z isn't better.
Also made it an option to set the AvgTime used by the powermeter, which is the time they use to average their readings over. 

Edited gradient descent settings so even less momentum . Set AvgTime here to 0.05 which seems to result in better optimization.